# api/openapi.yaml
openapi: 3.1.0
info:
  title: ML-Portal API
  version: 0.1.0
paths:
...
  /api/v1/admin/mode:
    post:
      tags:
      - admin
      - admin
      summary: Set Admin Mode
      description: Set maintenance/readonly mode
      operationId: set_admin_mode_api_v1_admin_mode_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Mode Data
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/models:
    get:
      tags:
      - admin
      - admin
      summary: List Models
      description: List models with filtering and pagination
      operationId: list_models_api_v1_admin_models_get
      parameters:
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by state
          title: State
        description: Filter by state
      - name: modality
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by modality
          title: Modality
        description: Filter by modality
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by model name
          title: Search
        description: Search by model name
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRegistryListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/models/{model_id}:
    get:
      tags:
      - admin
      - admin
      summary: Get Model Details
      description: Get detailed model information
      operationId: get_model_details_api_v1_admin_models__model_id__get
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - admin
      - admin
      summary: Update Model
      description: Update model state, default_for_new, or notes
      operationId: update_model_api_v1_admin_models__model_id__patch
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: Update Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      - admin
      summary: Delete Model
      description: Delete a model from registry
      operationId: delete_model_api_v1_admin_models__model_id__delete
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/models/{model_id}:retire:
    post:
      tags:
      - admin
      - admin
      summary: Retire Model
      description: Retire a model and optionally clean up related data
      operationId: retire_model_api_v1_admin_models__model_id__retire_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetireRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetireResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/models/{model_id}/tenants:
    get:
      tags:
      - admin
      - admin
      summary: Get Model Tenants
      description: Get tenants using a specific model
      operationId: get_model_tenants_api_v1_admin_models__model_id__tenants_get
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/rag/{doc_id}/models:
    get:
      tags:
      - rag
      - rag
      summary: Get Rag Document Models
      description: Get vectorized models for a document
      operationId: get_rag_document_models_api_v1_rag__doc_id__models_get
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tenants:
    get:
      tags:
      - tenants
      - tenants
      summary: List Tenants
      description: List tenants with pagination and filtering (admin only)
      operationId: list_tenants_api_v1_tenants_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: is_active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tenants
      - tenants
      summary: Create Tenant
      description: Create a new tenant (admin only)
      operationId: create_tenant_api_v1_tenants_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tenants/{tenant_id}:
    get:
      tags:
      - tenants
      - tenants
      summary: Get Tenant
      description: Get tenant by ID (admin only)
      operationId: get_tenant_api_v1_tenants__tenant_id__get
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - tenants
      - tenants
      summary: Update Tenant
      description: Update tenant (admin only)
      operationId: update_tenant_api_v1_tenants__tenant_id__put
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tenants
      - tenants
      summary: Delete Tenant
      description: Delete tenant (admin only)
      operationId: delete_tenant_api_v1_tenants__tenant_id__delete
      parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModelRegistry:
      properties:
        model:
          type: string
          title: Model
          description: Unique model identifier from manifest
        version:
          type: string
          title: Version
          description: Model version
        modality:
          type: string
          title: Modality
          description: 'Model modality: text|image|layout|table|rerank'
        state:
          type: string
          title: State
          description: 'Model state: active|archived|retired|disabled'
          default: active
        vector_dim:
          anyOf:
          - type: integer
          - type: 'null'
          title: Vector Dim
          description: Vector dimension for embedding models
        path:
          type: string
          title: Path
          description: Full path to model directory
        default_for_new:
          type: boolean
          title: Default For New
          description: Use as default for new tenants
          default: false
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Additional notes about the model
        id:
          type: string
          title: Id
        used_by_tenants:
          type: integer
          title: Used By Tenants
          description: Number of tenants using this model
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - model
      - version
      - modality
      - path
      - id
      - created_at
      - updated_at
      title: ModelRegistry
      description: Schema for model registry response
    ModelRegistryListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ModelRegistry'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: ModelRegistryListResponse
      description: Schema for paginated model registry list
...
    RetireRequest:
      properties:
        drop_vectors:
          type: boolean
          title: Drop Vectors
          description: Drop vector collections
          default: false
        remove_from_tenants:
          type: boolean
          title: Remove From Tenants
          description: Remove from tenant profiles
          default: false
      type: object
      title: RetireRequest
      description: Schema for model retirement request
    RetireResponse:
      properties:
        success:
          type: boolean
          title: Success
        affected_tenants:
          items:
            type: string
          type: array
          title: Affected Tenants
          description: Affected tenant IDs
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - message
      title: RetireResponse
      description: Schema for model retirement response
...
    Tenant:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: Tenant name
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
          description: Tenant description
        is_active:
          type: boolean
          title: Is Active
          description: Whether tenant is active
          default: true
        embed_models:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 2
          - type: 'null'
          title: Embed Models
          description: Embedding models (max 2)
        rerank_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Rerank Model
          description: Reranking model
        ocr:
          type: boolean
          title: Ocr
          description: Enable OCR processing
          default: false
        layout:
          type: boolean
          title: Layout
          description: Enable layout analysis
          default: false
        id:
          type: string
          title: Id
          description: Tenant ID
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: Last update timestamp
      type: object
      required:
      - name
      - id
      - created_at
      - updated_at
      title: Tenant
    TenantCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: Tenant name
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
          description: Tenant description
        is_active:
          type: boolean
          title: Is Active
          description: Whether tenant is active
          default: true
        embed_models:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 2
          - type: 'null'
          title: Embed Models
          description: Embedding models (max 2)
        rerank_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Rerank Model
          description: Reranking model
        ocr:
          type: boolean
          title: Ocr
          description: Enable OCR processing
          default: false
        layout:
          type: boolean
          title: Layout
          description: Enable layout analysis
          default: false
      type: object
      required:
      - name
      title: TenantCreate
    TenantListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Tenant'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: TenantListResponse
    TenantUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        embed_models:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 2
          - type: 'null'
          title: Embed Models
          description: Embedding models (max 2)
        rerank_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Rerank Model
          description: Reranking model
        ocr:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Ocr
          description: Enable OCR processing
        layout:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Layout
          description: Enable layout analysis
      type: object
      title: TenantUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError